impl: auto runtime upgrades
