--- docker/devnet/specs/pchain.genesis
@@ -1,1 +1,0 @@
-0x0000000000000000000000000000000000000000000000000000000000000000002cea5fa93713113dd2dbccf8647d6cc487e347341c44448429ff89cbd51e934203170a2e7597b7b7e3d84c05391d139a62b157e78786d8c082f29dcf4c11131400
