--- docker/devnet/specs/pchain1.key
@@ -1,7 +1,0 @@
-Secret phrase:       again increase pizza romance lyrics person giggle sketch toilet film volcano word
-  Network ID:        substrate
-  Secret seed:       0xe5d124d033b0ee38199bf3b8ad3dbd377ac5671ef3ca1103254d0f3094b88419
-  Public key (hex):  0x70728572b7e170831f6e0d495e40cb57e00bd91c5b36d067e1434ad78dc8b550
-  Account ID:        0x70728572b7e170831f6e0d495e40cb57e00bd91c5b36d067e1434ad78dc8b550
-  Public key (SS58): 5Ec9EtCUqwifGQibDPHrTQ5wZCEaF9odJy7JgkKWKFv7hq8K
-  SS58 Address:      5Ec9EtCUqwifGQibDPHrTQ5wZCEaF9odJy7JgkKWKFv7hq8K
