--- docker/devnet/specs/pchain2.key
@@ -1,7 +1,0 @@
-Secret phrase:       accident auto river zoo inner fitness employ occur bring weasel extra useful
-  Network ID:        substrate
-  Secret seed:       0x8df8af557c2d4bb7152ddf1a9a752fae8a4ce791fa6686207c2d52bdcdfc46e8
-  Public key (hex):  0xc8965cebe631da09cb96d11a211a132a6615758f00b42f3175d7609341a91b2e
-  Account ID:        0xc8965cebe631da09cb96d11a211a132a6615758f00b42f3175d7609341a91b2e
-  Public key (SS58): 5Gbi6pomDf9oeDJV3pXzAY8TCSsWanfTHhnfB7Qy6enMk4eZ
-  SS58 Address:      5Gbi6pomDf9oeDJV3pXzAY8TCSsWanfTHhnfB7Qy6enMk4eZ
