--- docker/devnet/specs/t3rn.genesis
@@ -1,1 +1,0 @@
-0x00000000000000000000000000000000000000000000000000000000000000000030ff968390226d03866a046599df0fae216881297618d67595b3b1c6244ae26d03170a2e7597b7b7e3d84c05391d139a62b157e78786d8c082f29dcf4c11131400
