--- docker/devnet/specs/t3rn1.key
@@ -1,7 +1,0 @@
-Secret phrase:       hundred crash oyster consider indicate acoustic resource empty cluster casual feed bundle
-  Network ID:        substrate
-  Secret seed:       0x2274654ead0b59545738f48f86bff07c3ae9877be45cd6f1099a38babb3272df
-  Public key (hex):  0x760e4371fce797f673dae4c51bfc6bedeb7b9f6abe83371b466f162c988f9759
-  Account ID:        0x760e4371fce797f673dae4c51bfc6bedeb7b9f6abe83371b466f162c988f9759
-  Public key (SS58): 5EjVk9P9V3FTDmhAJrEyYcpgHRUiLyYhYJLCYHT9ckJCmKx5
-  SS58 Address:      5EjVk9P9V3FTDmhAJrEyYcpgHRUiLyYhYJLCYHT9ckJCmKx5
