--- docker/devnet/specs/t3rn2.key
@@ -1,7 +1,0 @@
-Secret phrase:       arctic satoshi animal wheel reunion skull tornado oyster sniff pave clip jealous
-  Network ID:        substrate
-  Secret seed:       0xa3e048c0d80871debeb23fdf0d18f386cc576593d0428f186f2e06d7542e1e28
-  Public key (hex):  0xb0f1b025e82c2c587eb7e2657c41251e4e4caea75945e188faa7ef531744bb7a
-  Account ID:        0xb0f1b025e82c2c587eb7e2657c41251e4e4caea75945e188faa7ef531744bb7a
-  Public key (SS58): 5G4i5fVVx3nkefTM8Q1gpvYdqgPPTnPAosKBPNojkUCXMiGB
-  SS58 Address:      5G4i5fVVx3nkefTM8Q1gpvYdqgPPTnPAosKBPNojkUCXMiGB
